+++ draft-packages/collapsible/KaizenDraft/Collapsible/Collapsible.tsx
@@ -9,6 +9,8 @@ import { Sticky } from "./CollapsibleGroup"
 
 import styles from "./styles.scss"
 
+type Variant = "default" | "clear"
+
 export type Props = {
   id: string
   children: JSX.Element | JSX.Element[] | string
@@ -21,6 +23,13 @@ export type Props = {
   noSectionPadding?: boolean
   automationId?: string
   onToggle?: (open: boolean, id: string) => void
+
+  /**
+   * By default, the header will change background colour when open. When the variant
+    is set to 'clear', it will not have a background but a border-bottom will appear
+    to separate the heading from the content.
+   */
+  variant?: Variant
 
   /* Will avoid rendering the content until required (especially important when you have queries inside sections).
   Removes animation. */
@@ -51,15 +60,18 @@ class Collapsible extends React.Component<Props, State> {
       automationId,
       children,
       lazyLoad,
+      variant = "default",
     } = this.props
     const buttonId = `${this.props.id}-button`
     const sectionId = `${this.props.id}-section`
     const open = this.getOpen()
+    const isContainer = !group || separated
 
     return (
       <div
         className={classnames({
-          [styles.container]: !group || separated,
+          [styles.container]: isContainer,
+          [styles.stickyContainer]: isContainer && sticky,
           [styles.groupItem]: group && !separated,
           [styles.separated]: separated,
         })}
@@ -68,7 +80,8 @@ class Collapsible extends React.Component<Props, State> {
         <button
           id={buttonId}
           className={classnames(styles.button, {
-            [styles.open]: open,
+            [styles.defaultVariant]: open && variant === "default",
+            [styles.clearVariant]: open && variant === "clear",
             [styles.sticky]: sticky,
           })}
           style={sticky && { top: sticky.top }}