+++ draft-packages/modal/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.4.19](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.4.18...@kaizen/draft-modal@5.4.19) (2021-04-13)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [5.4.18](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.4.17...@kaizen/draft-modal@5.4.18) (2021-04-13)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [5.4.17](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.4.16...@kaizen/draft-modal@5.4.17) (2021-04-13)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [5.4.16](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.4.15...@kaizen/draft-modal@5.4.16) (2021-04-12)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [5.4.15](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.4.14...@kaizen/draft-modal@5.4.15) (2021-04-09)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [5.4.14](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.4.13...@kaizen/draft-modal@5.4.14) (2021-04-09)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [5.4.13](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.4.12...@kaizen/draft-modal@5.4.13) (2021-04-07)
+
+
+### Bug Fixes
+
+* React 17 fix - replace react-transition-group in draft-modal ([#1301](https://github.com/cultureamp/kaizen-design-system/issues/1301)) ([92899f7](https://github.com/cultureamp/kaizen-design-system/commit/92899f7a3155abf2ebb12075bcad7581ae8cf786))
+
+
+
+
+
+## [5.4.12](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.4.11...@kaizen/draft-modal@5.4.12) (2021-04-07)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
 ## [5.4.11](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.4.10...@kaizen/draft-modal@5.4.11) (2021-04-01)
 
 **Note:** Version bump only for package @kaizen/draft-modal