+++ draft-packages/search-box/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.36](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-search-box@1.1.35...@kaizen/draft-search-box@1.1.36) (2021-04-13)
+
+**Note:** Version bump only for package @kaizen/draft-search-box
+
+
+
+
+
+## [1.1.35](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-search-box@1.1.34...@kaizen/draft-search-box@1.1.35) (2021-04-13)
+
+**Note:** Version bump only for package @kaizen/draft-search-box
+
+
+
+
+
+## [1.1.34](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-search-box@1.1.33...@kaizen/draft-search-box@1.1.34) (2021-04-09)
+
+**Note:** Version bump only for package @kaizen/draft-search-box
+
+
+
+
+
+## [1.1.33](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-search-box@1.1.32...@kaizen/draft-search-box@1.1.33) (2021-04-09)
+
+**Note:** Version bump only for package @kaizen/draft-search-box
+
+
+
+
+
+## [1.1.32](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-search-box@1.1.31...@kaizen/draft-search-box@1.1.32) (2021-04-07)
+
+**Note:** Version bump only for package @kaizen/draft-search-box
+
+
+
+
+
+## [1.1.31](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-search-box@1.1.30...@kaizen/draft-search-box@1.1.31) (2021-04-07)
+
+**Note:** Version bump only for package @kaizen/draft-search-box
+
+
+
+
+
 ## [1.1.30](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-search-box@1.1.29...@kaizen/draft-search-box@1.1.30) (2021-03-31)
 
 **Note:** Version bump only for package @kaizen/draft-search-box