+++ draft-packages/split-button/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.10](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-split-button@2.3.9...@kaizen/draft-split-button@2.3.10) (2021-04-13)
+
+**Note:** Version bump only for package @kaizen/draft-split-button
+
+
+
+
+
+## [2.3.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-split-button@2.3.8...@kaizen/draft-split-button@2.3.9) (2021-04-07)
+
+**Note:** Version bump only for package @kaizen/draft-split-button
+
+
+
+
+
 ## [2.3.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-split-button@2.3.7...@kaizen/draft-split-button@2.3.8) (2021-03-31)
 
 **Note:** Version bump only for package @kaizen/draft-split-button