+++ draft-packages/tabs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.32](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tabs@4.1.31...@kaizen/draft-tabs@4.1.32) (2021-04-13)
+
+**Note:** Version bump only for package @kaizen/draft-tabs
+
+
+
+
+
+## [4.1.31](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tabs@4.1.30...@kaizen/draft-tabs@4.1.31) (2021-04-07)
+
+**Note:** Version bump only for package @kaizen/draft-tabs
+
+
+
+
+
 ## [4.1.30](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tabs@4.1.29...@kaizen/draft-tabs@4.1.30) (2021-03-31)
 
 **Note:** Version bump only for package @kaizen/draft-tabs