+++ packages/component-library/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.5.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@9.5.3...@kaizen/component-library@9.5.4) (2021-04-13)
+
+**Note:** Version bump only for package @kaizen/component-library
+
+
+
+
+
+## [9.5.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@9.5.2...@kaizen/component-library@9.5.3) (2021-04-07)
+
+**Note:** Version bump only for package @kaizen/component-library
+
+
+
+
+
 ## [9.5.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@9.5.1...@kaizen/component-library@9.5.2) (2021-03-31)
 
 **Note:** Version bump only for package @kaizen/component-library